adjusting visualize.py to tests for the south
