+++ tests/unit_tests/visualization/test_visualize.py
@@ -7,6 +7,7 @@ import astropy.units as u
 import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
 import numpy as np
+import pytest
 
 import simtools.util.general as gen
 from simtools.visualization import visualize
@@ -95,26 +96,36 @@ def test_add_unit():
     assert visualize._add_unit("Wavelength", value_without_unit) == "Wavelength"
 
 
-def test_get_telescope_patch(corsika_telescope_data_dict):
+def test_get_telescope_patch(manual_corsika_dict_north, manual_corsika_dict_south, io_handler):
+    def one_site(corsika_dict, x, y):
+        for tel_type in np.array(list(corsika_dict["corsika_sphere_radius"].keys())):
+            radius = corsika_dict["corsika_sphere_radius"][tel_type].value
+            patch = visualize.get_telescope_patch(tel_type, x, y, radius * u.m)
+            if mpatches.Circle == type(patch):
+                assert patch.radius == corsika_dict["corsika_sphere_radius"][tel_type].value
+            else:
+                assert isinstance(patch, mpatches.Rectangle)
 
-    for tel_type in np.array(list(corsika_telescope_data_dict["corsika_sphere_radius"].keys())):
-        radius = corsika_telescope_data_dict["corsika_sphere_radius"][tel_type].value
-        patch = visualize.get_telescope_patch(tel_type, 0 * u.m, 0 * u.m, radius * u.m)
-        if mpatches.Circle == type(patch):
-            assert (
-                patch.radius == corsika_telescope_data_dict["corsika_sphere_radius"][tel_type].value
-            )
-
-        else:
-            assert isinstance(patch, mpatches.Rectangle)
+    one_site(manual_corsika_dict_north, 0 * u.m, 0 * u.m)
+    one_site(manual_corsika_dict_south, 0 * u.m, 0 * u.m)
+    # Test passing other units
+    one_site(manual_corsika_dict_north, 0 * u.m, 0 * u.km)
+    one_site(manual_corsika_dict_south, 0 * u.cm, 0 * u.km)
+    with pytest.raises(TypeError):
+        one_site(manual_corsika_dict_south, 0, 0)
 
 
-def test_plot_array(telescope_north_test_file, layout_array_north_instance):
-    telescope_table = layout_array_north_instance.read_telescope_list_file(
-        telescope_north_test_file
-    )
-    telescopes_dict = layout_array_north_instance.include_radius_into_telescope_table(
-        telescope_table
-    )
-    fig_out = visualize.plot_array(telescopes_dict, rotate_angle=0 * u.deg)
-    assert isinstance(fig_out, type(plt.figure()))
+def test_plot_array(
+    telescope_north_test_file,
+    layout_array_north_instance,
+    telescope_south_test_file,
+    layout_array_south_instance,
+):
+    def one_site(test_file, instance):
+        telescope_table = instance.read_telescope_list_file(test_file)
+        telescopes_dict = instance.include_radius_into_telescope_table(telescope_table)
+        fig_out = visualize.plot_array(telescopes_dict, rotate_angle=0 * u.deg)
+        assert isinstance(fig_out, type(plt.figure()))
+
+    one_site(telescope_north_test_file, layout_array_north_instance)
+    one_site(telescope_south_test_file, layout_array_south_instance)
